requirements.txt
Adafruit-DHT==1.4.0
awscrt==0.16.17
AWSIoTPythonSDK==1.5.2
awsiotsdk==1.14.1
boto3==1.26.118
botocore==1.29.118
certifi==2020.6.20
chardet==4.0.0
colorzero==1.1
cupshelpers==1.0
dbus-python==1.2.16
distro==1.5.0
gpiozero==1.6.2
idna==2.10
jmespath==1.0.1
numpy==1.19.5
pexpect==4.8.0
picamera2==0.3.9
pidng==4.0.9
piexif==1.1.3
Pillow==8.1.2
pycairo==1.16.2
pycups==2.0.1
PyGObject==3.38.0
pysmbc==1.0.23
python-apt==2.2.1
python-dateutil==2.8.2
python-prctl==1.7
reportlab==3.5.59
requests==2.25.1
RPi.GPIO==0.7.0
s3transfer==0.6.0
simplejpeg==1.6.4
six==1.16.0
spidev==3.5
ssh-import-id==5.10
toml==0.10.1
urllib3==1.26.5
v4l2-python3==0.3.2
